+++ src/sim/utils/decide_position_algorithm/enum_algorithm.py
@@ -1,0 +1,5 @@
+from enum import IntEnum
+
+
+class Algorithm(IntEnum):
+    HIGH_POSITION = 1